requirements.txt
moviepy
praw
psaw
librosa
pillow 
selenium
python-dotenv
azure-cognitiveservices-speech